Do not use Turbo Drive for the `exportTheme` link (see #8048)
Description
-----------

### Description

* Fixes #8047 

This disables the prefetch of the download as mentioned in https://github.com/contao/contao/issues/8047#issuecomment-2646090568.

The current behavior tries to prefetch the download and Turbo tries to wait till we reach the destination URL (which we will never do). Disabling turbo is the way to go as we wouldn't want to download just by hovering over a link anyways.

Commits
-------

e1d3c7b2 Do not prefetch the theme download