Improve the language cache warmer performance (see #8123)
Description
-----------

In a project with a lot of extensions I noticed that the generation of the Contao language cache took __60 (!) seconds__. I was able to cut that time down to __30 seconds__ by using `contao.intl.enabled_locales: [en, de]`. However, I investigated further and found the following issue in our `ContaoCacheWarmer::generateLanguageCache()` method. I was able to further reduce it to __10 seconds__ with this PR.

We first search for all language files of a language via:

https://github.com/contao/contao/blob/7b080945f0a5fff05745a27cb188a098805a823f/core-bundle/src/Cache/ContaoCacheWarmer.php#L291-L295

Then we iterate over all language files and extract the filename - but then search for all language files _again_ - just with an additional filter on the name - via:

https://github.com/contao/contao/blob/7b080945f0a5fff05745a27cb188a098805a823f/core-bundle/src/Cache/ContaoCacheWarmer.php#L137-L141

This Finder search is done for all language file names (`default`, `tl_content`, `tl_news`, etc.) and thus the search time accumulates.

This PR filters the already available list of language files instead of searching for the files again. Even in a vanilla Contao 5.3 setup this cuts down the language cache generation from __15 seconds__ to just __4 seconds__ on my system.

Commits
-------

3bd29e6f Improve language cache warmer performance
fc67b140 use EmptyIterator in test instead
815ce062 Fix for PHP 8.1