Add a directory filter VFS decorator (see #8052)
Description
-----------

This PR adds a new decorator to the VFS (similar to the `PermissionCheckingVirtualFilesystem`) that supports a filtered view/access to the underlying storage.

When listing contents, you'll only see/access everything inside the directories that were specified as `$allowedDirectories`. Contrary to a prefixed "regular" VFS, you'll also see a root trail of directories. Concept-wise, these are virtual directories - when accessing these, the decorator will generate the data and not ask the underlying storage. 

Covered edge cases:
* If an allowed directory was specified but does exist in the underlying storage at this point, you can still write to it -Flysystem will create it on the fly (if needed by the adapter - otherwise, like on S3, it will just write the file). 
* Deleting allowed directories themselves is prevented.
* When providing allowed paths that overlap, only the top-level path is considered (because it covers the others).

Usage in the new filemanager will look something like this:
```php
$filterStorage = new DirectoryFilterVirtualFilesystem(
    new PermissionCheckingVirtualFilesystem(…),
    allowedDirectories: ['path', 'another/path', …]
);

// There is also a new helper method, that will list the first path containing "real" data
$initialViewPath = $filterStorage->getFirstNonVirtualDirectory();
```

Commits
-------

8e12edcb add a VFS decorator that will apply a filter on what can be accessed …
6e3753d2 fix phpstan annotation