Use separate signals to prevent executing connects/disconnects in the Choices controller (see #8203)
Description
-----------

Fixes #8192 

We need separate signals for when mutations are triggering subsequent connects/disconnects while adding or removing Choices elements. In the scenario in #8192 we are basically directly removing and readding an initialized Choices element in a single microtask. With the previous logic, the guard would still be up when readding, effectively stopping the connect

@zoglo Could you double check if this fixes the issue (and Choices is behaving correctly everywhere)?

Sidenote: I still dislike how Choices is altering the DOM (wrapping the element instead of putting its things aside) - maybe we should create an issue/PR there, to improve the situation. There are probably other frameworks (Vue? Anyone using mutation observers?) that are affected by this as well.

Commits
-------

4199b220 use separate signals for adding/removing