Make `Backend.modalSelector()` work from within a modal dialog (see #8226)
Description
-----------

A `dialog` element is always in the browsers `top` layer. Nothing can be displayed over it and - if opened as a modal - even setting `display: none` would not allow to pass any pointer events through. This is a good design for a modal window, but it would mean that our legacy `SimpleModal` overlay (which is not a dialog) would open behind an already open dialog.

I had this issue in the upcoming file manager, where I wanted to display the legacy edit mask in a dialog which contains URL pickers. 

This PR solves the issue by dispatching a custom event when the legacy modal is opened and reacting to it (suspend/resume) in the dialog component. (The changes also include a minor CSS improvement to the dialog component, so that arbitrary buttons are rendered correctly.)

Commits
-------

0c4c93d4 allow simple modal to be displayed "over" dialog
32080b2c build assets