Rewrite the SERP widget to a Stimulus controller (see #8257)
Description
-----------

This PR introduces a new `contao--serp-preview` controller and removes its MooTools counterpart. In the process I factored out the HTML into a template and simplified the code. Before, the initial values were generated with PHP, now the Stimulus controller handles that as well removing the risk of inconsistencies.

In order to decouple things, the `contao--tinymce` controller now updates the underlying textarea element on `keydown` in the editor and dispatches an `input` event (also on the textarea element). This way other code can listen on the input field like with a native textarea and there is no need to use the controller in a Stimulus outlet/wait for its initialization/use the global tinyMCE instance (dangerous)/use tinyMCE-specific events.

Commits
-------

83063baf make serp widget a Stimulus controller
059b7775 remove moo version
68e9b1aa Make the error message configurable in the template

Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>