Fix page sorting with same languages across domains (see #8270)
Description
-----------

Today we had a very edge case on a client system with incorrect sorting of pages / incorrect match in the front end. I was able to reproduce and write a test case including anonymized data. The test data is pretty extensive, but I removed what was possible without causing the issue to be gone.

Basically, I think the page results were not fully sorted because it did not sort the same language across different domains.

Commits
-------

7ca6e45d Fix page sorting with same languages across domains