Use a wrapper for Choices.js (see #8360)
Description
-----------

Instead of reinventing accessibility for another select script or custom shadow dom handling, this simply adds a wrapper DIV to solve the Choices initialization problem. It works 99% out of the box. There might be small CSS issues still in the back end, not sure I catched everything, but they sure are easy to solve.

Commits
-------

ab919306 Use a wrapper for Choices.js
5c8a5091 Fix closing div
1fd5b6dd Improve wizards
30c38969 Rebuild assets
5695869e Properly set spacings for `tl_select_wrapper` and its contents
5f07ab15 Expand `.tl_select_wrapper` on mobile devices
d2074838 Fix the unit tests and rebuild the assets

Co-authored-by: zoglo <55794780+zoglo@users.noreply.github.com>
Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>