Add a helper for the searchable content on the search document (see #8370)
Description
-----------

This extracts some of the logic that is currently hardcoded in `Search.php` and thus would have to be repeated by every search indexer.

I tried re-using this logic in the `Search.php` but it doesn't really work since there's also an `indexPage` hook that would then get different arguments etc. so I think we should just leave that as is.

Commits
-------

5b5c8c49 Add a helper for the searchable content on the search document