Fix recursive paste permissions (see #8277)
Description
-----------

Fixes https://github.com/contao/contao/issues/8264. The issue was is that `$blnCircularReference` is true for any clipboard mode (copy or paste), because it "represents" whether the current ID is a child of the current clipboard, not whether pasting it would "lead to a circular problem". As @zoglo noticed in the issue, the cut-check was missing.

This also fixes two cases of incorrectly checking the clipboard if multiple IDs are copied/moved. Both cases are easily reproducable in the online demo:

 1. when limiting the tree to a subnode, and using edit-all to move the root node, the paste buttons incorrectly enabled

    https://github.com/user-attachments/assets/377e496f-a9ac-477d-b06c-546dcdf56386

2. when copying multiple records and ajax-loading a sub-sub-part of the tree, the paste buttons are incorrectly enabled

    https://github.com/user-attachments/assets/b7fb354c-1f75-4a80-9449-db9342a24ba5


These bugs likely also happen in Contao 5.3, but the fix would be completely different and not sure they are worth it.

Commits
-------

b6e815d9 Fix recursive paste permissions
a43cf98f Use the correct table variable
911f8011 Use the correct table variable
22003365 Use the correct table variable
6e2fb903 Fix strong comparison of clipboard ID