Add a shortcut to get the `HtmlHeadBag` (see #8393)
Description
-----------

I find it rather cumbersome at the moment to add a title or other data to the `HtmlHeadBag`. I need to DI the response context, get the current one, test for whether the HtmlHeadBag is available and then add something.

This shortcut would allow (fragment) controllers to do something as simple as `$this->getHtmlHeadBag()?->setTitle(...)`. I'm aware there can be other bags in the response context, but afaik there currently are none that are necessary in a (fragment) controller.

Commits
-------

233c4f38 Add a shortcut to get the HtmlHeadBag
e75ba63d Make the service optional

Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>