Fix absolute path URLs not enabling the lightbox (see #8379)
Description
-----------

If you have an insert tag reference in the `link` metadata of a file, the lightbox will not be enabled even though the target is an image. This worked in Contao 4 - but broke in Contao 5 due to the insert tag now returning an absolute URL path.

1. In the file manager edit an image.
2. In its metadata, click the wizard to insert something into the _Link_ field.
3. Go to _Files_ and set a different image as the target.
4. Create a new `image` content element and select the former image.
5. Enable **Full-size view/new window** and save.

In the front end in Contao 4.13, the link will have `data-lightbox`. In Contao 5 the link will not have `data-lightbox`. See also the adjusted test case.

The issue is that in our `$getResourceOrUrl` we check for `Path::isAbsolute()`. However, this will also be true for path absolute URLs - and in Contao 5 all such references are path absolute URLs now (and not relative to the base href anymore).

This PR changes the implementation by checking whether the `$target` references a file relative to the project dir or relative to the public dir.

It also keeps the check whether it already references an absolute filesystem path - though this should never actually be the case.

Commits
-------

f0c8e0ed Fix file insert tag references not enabling lightbox
f092f984 expand check to all absolute URL paths
6a64c48b also decode
1c86eca3 overhaul getResourceOrUrl
990a968b Correctly return URL for public dir file
6dd8f68d update test
1713775f do not decode filesystem absolute path