Use `Backend::addToUrl()` when switching to "edit multiple" mode (see #8517)
Description
-----------

In 5.x-dev, if you select multiple records an click on `Edit`, you will get the request token warning. That's because `act=select` is prefetched (= has no request token), but the `editAll` and other actions require a request token.

Technically, `prefetch = true` was added in 5.5 on the `all` operation, but it did not work. This was fixed with https://github.com/contao/contao/pull/8346 which caused this issue.

Commits
-------

1e957be7 Use Backend::addToUrl when switching edit-all mode