Always add a close button to messages (see #8483)
Description
-----------

Current, if a popup message is successful, it will automatically expires after 5000 seconds, but it will not have a close button. I think this is bad UX, regardless of whether the message disappears automatically, I still want to be able to close it faster.

This also removes the unnecessary `contao.backend_user` variable, which has a wrong method name (see https://github.com/contao/contao/pull/8472#issuecomment-3011929978) and the global Config property, which is only there for backwards compatibility for other properties that were previously global and not user-specific.

Commits
-------

bb62a6ca Always add close button to messages
29aecb5a Tests
14ba24ca Add the autoclose condition