Fix the operation menu position in the "paste into" view (see #8537)
Description
-----------

### Description

Fixes #8536 and is caused by two things:

As already outlined in https://github.com/contao/contao/pull/8518#discussion_r2185350941, calculating the offset from the button was a little bit arbitrary so it was changed to offset parent. However, offsetParent does not work if the ancestor or the element itself are set to `display: none`. Also using the buttonTarget prior to b0244f11b0bdc4b22223fd2bff6e7892aa52021a may not work as it's not existing in the dom and not getting a proper boundary due to being `display: none`.

1. This PR rewrites it to get the parentRect by going top-down instead of bottom-up (as the `menuTarget` and `operations-menu-container` will always exist)

2. Removing the `invisible` class that adds a `overflow: hidden`. It's actually not needed in this case as everything is hidden and the screenreaders do not need to read anything within a hidden menu that can't be targeted / focused.

@aschempp can you confirm this fixes your issue?

Commits
-------

19ae8d4f Remove the `invisible` class
650ef8cf Get the `parentRect` from the parent