Deprecate `Backend.enableImageSizeWidgets()` (see #8242)
Description
-----------

I noticed that we are still using `Backend.enableImageSizeWidgets()` even though we have a proper controller for the image size widget since #5610 (and further improved in #8075). This PR deprecates it and moves the functionality to the stimulus controller instead.

Commits
-------

118d5e82 Deprecate Backend.enableImageSizeWidgets()
e2a7517d Merge with 5.x
0522a4a8 Optimize split, remove deprecation, use proper targets
c945064f Fix inputTargetDisconnected